tests/wpt/web-platform-tests/css/css-inline/parsing/baseline-shift-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Inline Layout: parsing baseline-shift with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-inline-3/#baseline-shift-property">
<meta name="assert" content="baseline-shift supports only the grammar 'length-percentage | sub | super'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
// Valid in SVG 1.1. Accepted by Blink and WebKit.
// https://github.com/w3c/csswg-drafts/issues/3975
test_invalid_value("baseline-shift", "5");

test_invalid_value("baseline-shift", "-10px 20%");
test_invalid_value("baseline-shift", "calc(2em + 3ex), 0");
test_invalid_value("baseline-shift", "sub super");
test_invalid_value("baseline-shift", "20px sub");
</script>
</body>
</html>
